Refactored icon matching into components
