+++ sto_sister/stages/icon_matching.py
@@ -1,6 +1,7 @@
 from typing import Any, Callable, Dict, List, Tuple, Optional
 
 from ..pipeline import Stage, StageResult, PipelineContext
+from ..utils.image import apply_mask, load_quality_overlays, show_image
 from ..components.icon_matcher import IconMatcher
 
 class IconMatchingStage(Stage):
@@ -10,17 +11,16 @@ class IconMatchingStage(Stage):
         super().__init__(opts, app_config)
 
         self.matcher = IconMatcher(
-            hash_index=app_config.get("hash_index"),
             debug=opts.get("debug", False),
-            engine_type=opts.get("engine_type", "ssim"),
         )
 
     def run(
         self, ctx: PipelineContext, report: Callable[[str, float], None]
     ) -> StageResult:
         report(self.name, 0.0)
+        
         icon_sets = ctx.app_config.get("icon_sets", {})
-        ctx.overlays = self.matcher.load_quality_overlays(
+        ctx.overlays = load_quality_overlays(
             ctx.config.get("overlay_dir", "")
         )
         # print(f"[Matching] ctx.filtered_icons: {ctx.filtered_icons}")
